+++ libqpdf/qpdf/BitWriter.hh
@@ -15,7 +15,7 @@ class BitWriter
     QPDF_DLL
     BitWriter(Pipeline* pl);
     QPDF_DLL
-    void writeBits(unsigned long val, unsigned int bits);
+    void writeBits(unsigned long long val, unsigned int bits);
     // Force any partial byte to be written to the pipeline.
     QPDF_DLL
     void flush();
